Add user action helper
